refactor date formatting to reusable util
